+++ src/interfaces/plugins/liveSearch/editLiveSearch.interface.ts
@@ -11,4 +11,14 @@ export interface EditLiveSearchCssClasses extends LiveSearchCssClasses
      * Applied to HTML element that represents the search input
      */
     searchElement: string;
+
+    /**
+     * Applied to HTML element that represents the search input when the select is in multiple mode
+     */
+    multiple: string;
+
+    /**
+     * Applied to HTML element that represents the search input when the select is in single mode
+     */
+    single: string;
 }