+++ src/pipes/index.ts
@@ -3,3 +3,4 @@ export * from './displayValue/displayValue.pipe';
 export * from './getPlugin/getPlugin.pipe';
 export * from './groupedListOptions/groupedListOptions.pipe';
 export * from './hasValue/hasValue.pipe';
+export * from './optionCssClasses/optionCssClasses.pipe';